requirements-core.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-multipart>=0.0.9
pillow>=10.0.0
opencv-python-headless>=4.9.0
numpy>=1.26.0
openai>=1.40.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
httpx>=0.27.0
